refactor(ProductUiModel): 패키지 이동
