+++ app/src/main/java/woowacourse/shopping/database/ShoppingRepository.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.database
 
-import woowacourse.shopping.productdetail.ProductUiModel
+import woowacourse.shopping.model.ProductUiModel
 
 interface ShoppingRepository {
