+++ app/src/main/java/woowacourse/shopping/database/product/MockProduct.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.database.product
 
-import woowacourse.shopping.productdetail.ProductUiModel
+import woowacourse.shopping.model.ProductUiModel
 
 object MockProduct {
