+++ app/src/main/java/woowacourse/shopping/shopping/ShoppingContract.kt
@@ -1,6 +1,6 @@
 package woowacourse.shopping.shopping
 
-import woowacourse.shopping.productdetail.ProductUiModel
+import woowacourse.shopping.model.ProductUiModel
 
 interface ShoppingContract {
