+++ app/src/main/java/woowacourse/shopping/shopping/ShoppingRecyclerAdapter.kt
@@ -3,7 +3,7 @@ package woowacourse.shopping.shopping
 import android.annotation.SuppressLint
 import android.view.ViewGroup
 import androidx.recyclerview.widget.RecyclerView
-import woowacourse.shopping.productdetail.ProductUiModel
+import woowacourse.shopping.model.ProductUiModel
 
 class ShoppingRecyclerAdapter(
     private val products: List<ProductUiModel>,
