+++ app/src/test/java/woowacourse/shopping/ShoppingCartPresenterTest.kt
@@ -10,7 +10,7 @@ import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Test
 import woowacourse.shopping.database.ShoppingRepository
-import woowacourse.shopping.productdetail.ProductUiModel
+import woowacourse.shopping.model.ProductUiModel
 import woowacourse.shopping.shoppingcart.ShoppingCartContract
 import woowacourse.shopping.shoppingcart.ShoppingCartPresenter
